+++ UnstoppableWallet/UnstoppableWallet/Modules/Transactions/RecordServices/SingleWalletRecordService.swift
@@ -54,7 +54,7 @@ extension SingleWalletRecordService: ITransactionRecordService {
                 .disposed(by: disposeBag)
     }
 
-    func set(typeFilter: TransactionsModule2.TypeFilter) {
+    func set(typeFilter: TransactionTypeFilter) {
         dataSource.set(typeFilter: typeFilter)
     }
