--- UnstoppableWallet/UnstoppableWallet/Modules/Transactions/TransactionViewStatus.swift
@@ -1,4 +1,0 @@
-struct TransactionViewStatus {
-    let showProgress: Bool
-    let showMessage: Bool
-}
